refactor(config): change timeout type from Long to Integer
거래소 API 설정에서 타임아웃 타입을 통일합니다.
HTTP 타임아웃 설정은 일반적으로 Integer로 사용되며,
Spring Boot의 표준 타임아웃 설정과 일치시킵니다.

- BinanceProperties: connectionTimeout, readTimeout Long → Integer
- CoinoneProperties: connectionTimeout, readTimeout Long → Integer

+++ src/main/java/sungjun/bitcoin/algorithmtrading/infrastructure/config/coinone/CoinoneProperties.java
@@ -76,13 +76,13 @@ public class CoinoneProperties {
      * HTTP 연결 타임아웃 (초 단위)
      */
     @NotNull(message = "연결 타임아웃은 필수 값입니다.")
-    private Long connectionTimeout;
+    private Integer connectionTimeout;
 
     /**
      * HTTP 읽기 타임아웃 (초 단위)
      */
     @NotNull(message = "읽기 타임아웃은 필수 값입니다.")
-    private Long readTimeout;
+    private Integer readTimeout;
 
     /**
      * 서명 생성에 사용할 알고리즘 (예: HmacSHA512)
@@ -102,7 +102,7 @@ public class CoinoneProperties {
      * @param signatureAlgorithm the algorithm name used for signature generation (e.g., "HmacSHA512")
      */
     @ConstructorBinding
-    public CoinoneProperties(String publicUrl, String privateUrl, String accessToken, String secretKey, Long connectionTimeout, Long readTimeout, String signatureAlgorithm) {
+    public CoinoneProperties(String publicUrl, String privateUrl, String accessToken, String secretKey, Integer connectionTimeout, Integer readTimeout, String signatureAlgorithm) {
         this.publicUrl = publicUrl;
         this.privateUrl = privateUrl;
         this.accessToken = accessToken;
